+++ src/main/scala/com/github/gchudnov/mtg/diagram/internal/BasicTranslator.scala
@@ -7,7 +7,7 @@ import com.github.gchudnov.mtg.Interval
 import com.github.gchudnov.mtg.Diagram.Canvas
 import com.github.gchudnov.mtg.Diagram
 import com.github.gchudnov.mtg.Diagram.Theme
-import com.github.gchudnov.mtg.Diagram.Translator
+import com.github.gchudnov.mtg.diagram.Translator
 import com.github.gchudnov.mtg.diagram.Span
 import com.github.gchudnov.mtg.Diagram.View
 import com.github.gchudnov.mtg.Diagram.Tick
